// src/modules/users/presentation/models/UserFormData.ts
export interface UserFormData {
  username: string;
  email: string;
  password?: string;
  confirmPassword?: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  sexId: string;
  idCard: string;
  citizenId: string;
  positionId: string;
  contractTypeId: string;
  employeeStatusId: string;
  hireDate: string;
  terminationDate: string;
  baseSalary: string;
  supervisorId: string;
  assignedZones: string;
  driverLicense: string;
  hasCompanyVehicle: string;
  internalPhone: string;
  internalEmail: string;
  photoUrl: string;
  createdBy: string;
}